Set up custom chai assertion; curriculum scripts pass
